:recycle: extract function to utility
